Fix handling of InterruptedExceptions
These exceptions were handled incorrectly throughout the whole
codebase, usually rethrown without restoring the interrupt status
or discarded entirely. This means that the system would not stop
executing even though the user wanted it to. In some cases this
also left the system in an inconsistent state, leading to deadlocks.

The most notable changes include:

- UncheckedException.rethrow automatically restores the interrupt status
- AsyncDispatch is guaranteed to deliver its messages, even when interrupted
- ExecHandle cancels the started process if it is interrupted while waiting
- ExecHandle disconnects from the process' output before killing it
- The worker API cancels the started work items if it is interrupted
- ManagedExecutors shut down immediately if they are interrupted while stopping
- We no longer log exceptions caused by interruption to the console
- Interrupting our caches no longer leaks file locks